[4.4] Move the Cypress Tests to ESM
